# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing experience data collection",
        "title": "Formbricks Hub API",
        "contact": {
            "name": "Tiago",
            "email": "xxxxx@xxxxx.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/health": {
            "get": {
                "description": "Check if the API is running",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/experiences": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of experience data records with optional filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiences"
                ],
                "summary": "List experience data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by source type",
                        "name": "source_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source ID",
                        "name": "source_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by field ID",
                        "name": "field_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user identifier",
                        "name": "user_identifier",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum number of records to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of records to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ExperienceData"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new experience data record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiences"
                ],
                "summary": "Create experience data",
                "parameters": [
                    {
                        "description": "Experience data to create",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateExperienceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ExperienceData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/experiences/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search experience data with advanced filters, full-text search, and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiences"
                ],
                "summary": "Search experience data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Full-text search query",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source type",
                        "name": "source_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by source ID",
                        "name": "source_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by field ID",
                        "name": "field_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by field type",
                        "name": "field_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user identifier",
                        "name": "user_identifier",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by collected_at \u003e= start_date (RFC3339 format)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by collected_at \u003c= end_date (RFC3339 format)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results per page (default 20, max 40)",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (starts at 0, default 0)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SearchExperiencesResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/experiences/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single experience data record by its UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiences"
                ],
                "summary": "Get experience data by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Experience ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExperienceData"
                        }
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Experience not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an experience data record by ID",
                "tags": [
                    "experiences"
                ],
                "summary": "Delete experience data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Experience ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content - Successfully deleted"
                    },
                    "400": {
                        "description": "Invalid UUID format",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Experience not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing experience data record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "experiences"
                ],
                "summary": "Update experience data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Experience ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateExperienceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ExperienceData"
                        }
                    },
                    "400": {
                        "description": "Invalid request or UUID format",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing API key",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Experience not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.CreateExperienceRequest": {
            "type": "object",
            "properties": {
                "collected_at": {
                    "type": "string"
                },
                "field_id": {
                    "type": "string"
                },
                "field_label": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object"
                },
                "source_id": {
                    "type": "string"
                },
                "source_name": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "user_identifier": {
                    "type": "string"
                },
                "value_boolean": {
                    "type": "boolean"
                },
                "value_date": {
                    "type": "string"
                },
                "value_json": {
                    "type": "object"
                },
                "value_number": {
                    "type": "number"
                },
                "value_text": {
                    "type": "string"
                }
            }
        },
        "models.ExperienceData": {
            "type": "object",
            "properties": {
                "collected_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "field_id": {
                    "type": "string"
                },
                "field_label": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object"
                },
                "source_id": {
                    "type": "string"
                },
                "source_name": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_identifier": {
                    "type": "string"
                },
                "value_boolean": {
                    "type": "boolean"
                },
                "value_date": {
                    "type": "string"
                },
                "value_json": {
                    "type": "object"
                },
                "value_number": {
                    "type": "number"
                },
                "value_text": {
                    "type": "string"
                }
            }
        },
        "models.SearchExperiencesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExperienceData"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "models.UpdateExperienceRequest": {
            "type": "object",
            "properties": {
                "field_id": {
                    "type": "string"
                },
                "field_label": {
                    "type": "string"
                },
                "field_type": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object"
                },
                "source_id": {
                    "type": "string"
                },
                "source_name": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "user_identifier": {
                    "type": "string"
                },
                "value_boolean": {
                    "type": "boolean"
                },
                "value_date": {
                    "type": "string"
                },
                "value_json": {
                    "type": "object"
                },
                "value_number": {
                    "type": "number"
                },
                "value_text": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your API key.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}